Parse command-line options for pumactl (#1482)
* Parse command-line options for pumactl

Before, we would remove any non-yielding
args, but missed to actually parse them.
Resulting in issues like, --pidfile or
similar args not taking any effect.

* Remove misleading message of puma starting on pumactl exit

This way, when puma shuts down (running from pumactl), it does not end
with "Puma is started" message. Because, the copy is dependent
on the value of `@command`, the same will always point
to `start`, if pumactl was issued a `start`, hence during
exit, this message appeared as misleading.