Remove broken puma/convenient [changelog skip] (#2039)
This file was added in 2016 in 663666c5, then broken in 2017 in
cda9317e. `c` was renamed to `user_config`, but one reference to `c` was
missed.

To confirm the method is indeed broken I ran:

```rb
require 'puma/convenient'
Puma.run
```

Since this has broken for more than 2 years, I think it is pretty safe
to say nobody is using it. That is why I propose removing instead of
fixing it.