Clarify how to run puma in single mode
When starting to transition from Unicorn to Puma, I first tried out
puma in development. I initially assumed from skimming the docs that
the number of workers for single mode should be 1. This led to our
developers not being able to kill their puma server properly in our
dev environment via Ctrl+C since the master process would stick
around and try to revive it.

Upon reading the docs more carefully and checking out the issues,
I found [I wasn't alone][issue-on-single-mode]. While it's obvious
that 1 worker implies the existence of a master process, when you're
just starting out it might take you a little while before you get
your head around single mode versus cluster mode and having the
configuration clearly stated helps with that imo.

So to make it blatantly obvious how to run puma in single mode,
I've added a line to the docs to explain it.

[issue-on-single-mode]: https://github.com/puma/puma/issues/1364