add support for the PROXY protocol (v1 only) (#2654)
* add support for the PROXY protocol (v1 only)

* slightly simplify test

* address review feedback

* move PROXY protocol parsing into its own method; avoid issue with short reads

The HTTP parser requires that the buffer be non-empty when it's invoked,
but if the entire buffer was the PROXY protocol, we may sometimes invoke
it with an empty buffer. This causes the following error:

  Puma::HttpParserError: Requested start is after data buffer end.

The fix? Any time we consume the entire buffer for the PROXY protocol,
simply return `false` to indicate that we require more data.